furbaby/requirements.txt
asgiref==3.7.2
astroid==2.15.8
attrs==22.2.0
awsebcli==3.20.10
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
black==23.9.1
bleach==6.0.0
blessed==1.20.0
botocore==1.31.64
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.3
coverage==6.5.0
coveralls==3.3.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.0
django-cors-headers==4.2.0
django-heartbeat==2.2.0
djangorestframework==3.14.0
docopt==0.6.2
drf-standardized-errors==0.12.5
exceptiongroup==1.1.3
fastjsonschema==2.16.2
filelock==3.12.4
flake8==5.0.4
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter-client==8.0.2
jupyter-core==5.2.0
jupyterlab-pygments==0.2.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mistune==2.0.5
mypy-extensions==1.0.0
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
packaging==23.2
pandocfilters==1.5.0
pathspec==0.10.1
pgsanity==0.2.9
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
psutil==5.9.5
psycopg==3.1.12
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.14.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyrsistent==0.19.3
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.0.0
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
soupsieve==2.4
sqlparse==0.4.4
termcolor==1.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.2
traitlets==5.9.0
typing-extensions==4.8.0
urllib3==1.26.17
virtualenv==20.24.5
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.15.0
zipp==3.13.0
